+++ test/unittest/CMakeLists.txt
@@ -23,7 +23,7 @@ add_subdirectory(dds/status)
 add_subdirectory(dds/subscriber)
 add_subdirectory(dds/topic)
 add_subdirectory(dds/topic/DDSSQLFilter)
-add_subdirectory(dds/xtypes/dyndata_idl)
+add_subdirectory(dds/xtypes/dyndata_json)
 add_subdirectory(dds/xtypes/type_representation)
 add_subdirectory(logging)
 add_subdirectory(rtps/attributes)
