--- test/unittest/dds/xtypes/dyndata_idl/types/data_test/common/DataTestCommon.cpp
@@ -1,31 +1,0 @@
-// Copyright 2022 Proyectos y Sistemas de Mantenimiento SL (eProsima).
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/**
- * @file DataTestCommon.cpp
- *
- */
-
-#include <fastdds/dds/xtypes/dynamic_types/DynamicData.hpp>
-#include <fastdds/dds/xtypes/dynamic_types/DynamicDataFactory.hpp>
-
-#include "../../types.hpp"
-#include "../gen/DataTest.hpp"
-#include "../gen/DataTestTypeObjectSupport.hpp"
-
-template <>
-void register_type_object_representation_gen<DataTypeKind::DATA_TEST>()
-{
-    register_DataTest_type_objects();
-}
