+++ test/unittest/dds/xtypes/dyndata_json/types/comprehensive_type/ComprehensiveType_EPROSIMA.hpp
@@ -1,0 +1,374 @@
+// Copyright 2024 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <string>
+
+std::string expected_json_comprehensive_eprosima = R"({
+    "complex_array": [
+        {
+            "bitmask_sequence": [],
+            "complex_union": {
+                "fourth": {
+                    "second": 0
+                }
+            },
+            "inner_union": {
+                "second": 0
+            },
+            "long_array": [
+                [
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ],
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ],
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ]
+                ],
+                [
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ],
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ],
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ]
+                ]
+            ],
+            "my_aliased_bounded_string": "",
+            "my_aliased_enum": {
+                "name": "A",
+                "value": 0
+            },
+            "my_aliased_struct": {
+                "my_bool": false,
+                "my_char": "\u0000",
+                "my_double": 0.0,
+                "my_float": 0.0,
+                "my_int8": 0,
+                "my_long": 0,
+                "my_longdouble": 0.0,
+                "my_longlong": 0,
+                "my_octet": 0,
+                "my_short": 0,
+                "my_uint8": 0,
+                "my_ulong": 0,
+                "my_ulonglong": 0,
+                "my_ushort": 0,
+                "my_wchar": "\u0000"
+            },
+            "my_bitmask": {
+                "active": [],
+                "binary": "00000000000000000000000000000000",
+                "value": 0
+            },
+            "my_bitset": {
+                "a": 0,
+                "b": false,
+                "c": 0,
+                "d": 0
+            },
+            "my_bool": false,
+            "my_bounded_string": "",
+            "my_bounded_wstring": "",
+            "my_char": "\u0000",
+            "my_double": 0.0,
+            "my_enum": {
+                "name": "A",
+                "value": 0
+            },
+            "my_float": 0.0,
+            "my_int8": 0,
+            "my_long": 0,
+            "my_longdouble": 0.0,
+            "my_longlong": 0,
+            "my_octet": 0,
+            "my_recursive_alias": {
+                "name": "A",
+                "value": 0
+            },
+            "my_short": 0,
+            "my_string": "",
+            "my_uint8": 0,
+            "my_ulong": 0,
+            "my_ulonglong": 0,
+            "my_ushort": 0,
+            "my_wchar": "\u0000",
+            "my_wstring": "",
+            "short_long_map": null,
+            "short_sequence": [],
+            "string_alias_unbounded_map": null,
+            "string_unbounded_map": null
+        },
+        {
+            "bitmask_sequence": [],
+            "complex_union": {
+                "fourth": {
+                    "second": 0
+                }
+            },
+            "inner_union": {
+                "second": 0
+            },
+            "long_array": [
+                [
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ],
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ],
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ]
+                ],
+                [
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ],
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ],
+                    [
+                        0,
+                        0,
+                        0,
+                        0
+                    ]
+                ]
+            ],
+            "my_aliased_bounded_string": "",
+            "my_aliased_enum": {
+                "name": "A",
+                "value": 0
+            },
+            "my_aliased_struct": {
+                "my_bool": false,
+                "my_char": "\u0000",
+                "my_double": 0.0,
+                "my_float": 0.0,
+                "my_int8": 0,
+                "my_long": 0,
+                "my_longdouble": 0.0,
+                "my_longlong": 0,
+                "my_octet": 0,
+                "my_short": 0,
+                "my_uint8": 0,
+                "my_ulong": 0,
+                "my_ulonglong": 0,
+                "my_ushort": 0,
+                "my_wchar": "\u0000"
+            },
+            "my_bitmask": {
+                "active": [],
+                "binary": "00000000000000000000000000000000",
+                "value": 0
+            },
+            "my_bitset": {
+                "a": 0,
+                "b": false,
+                "c": 0,
+                "d": 0
+            },
+            "my_bool": false,
+            "my_bounded_string": "",
+            "my_bounded_wstring": "",
+            "my_char": "\u0000",
+            "my_double": 0.0,
+            "my_enum": {
+                "name": "A",
+                "value": 0
+            },
+            "my_float": 0.0,
+            "my_int8": 0,
+            "my_long": 0,
+            "my_longdouble": 0.0,
+            "my_longlong": 0,
+            "my_octet": 0,
+            "my_recursive_alias": {
+                "name": "A",
+                "value": 0
+            },
+            "my_short": 0,
+            "my_string": "",
+            "my_uint8": 0,
+            "my_ulong": 0,
+            "my_ulonglong": 0,
+            "my_ushort": 0,
+            "my_wchar": "\u0000",
+            "my_wstring": "",
+            "short_long_map": null,
+            "short_sequence": [],
+            "string_alias_unbounded_map": null,
+            "string_unbounded_map": null
+        }
+    ],
+    "complex_map": null,
+    "complex_sequence": [],
+    "index": 0,
+    "inner_struct": {
+        "bitmask_sequence": [],
+        "complex_union": {
+            "fourth": {
+                "second": 0
+            }
+        },
+        "inner_union": {
+            "second": 0
+        },
+        "long_array": [
+            [
+                [
+                    0,
+                    0,
+                    0,
+                    0
+                ],
+                [
+                    0,
+                    0,
+                    0,
+                    0
+                ],
+                [
+                    0,
+                    0,
+                    0,
+                    0
+                ]
+            ],
+            [
+                [
+                    0,
+                    0,
+                    0,
+                    0
+                ],
+                [
+                    0,
+                    0,
+                    0,
+                    0
+                ],
+                [
+                    0,
+                    0,
+                    0,
+                    0
+                ]
+            ]
+        ],
+        "my_aliased_bounded_string": "",
+        "my_aliased_enum": {
+            "name": "A",
+            "value": 0
+        },
+        "my_aliased_struct": {
+            "my_bool": false,
+            "my_char": "\u0000",
+            "my_double": 0.0,
+            "my_float": 0.0,
+            "my_int8": 0,
+            "my_long": 0,
+            "my_longdouble": 0.0,
+            "my_longlong": 0,
+            "my_octet": 0,
+            "my_short": 0,
+            "my_uint8": 0,
+            "my_ulong": 0,
+            "my_ulonglong": 0,
+            "my_ushort": 0,
+            "my_wchar": "\u0000"
+        },
+        "my_bitmask": {
+            "active": [],
+            "binary": "00000000000000000000000000000000",
+            "value": 0
+        },
+        "my_bitset": {
+            "a": 0,
+            "b": false,
+            "c": 0,
+            "d": 0
+        },
+        "my_bool": false,
+        "my_bounded_string": "",
+        "my_bounded_wstring": "",
+        "my_char": "\u0000",
+        "my_double": 0.0,
+        "my_enum": {
+            "name": "A",
+            "value": 0
+        },
+        "my_float": 0.0,
+        "my_int8": 0,
+        "my_long": 0,
+        "my_longdouble": 0.0,
+        "my_longlong": 0,
+        "my_octet": 0,
+        "my_recursive_alias": {
+            "name": "A",
+            "value": 0
+        },
+        "my_short": 0,
+        "my_string": "",
+        "my_uint8": 0,
+        "my_ulong": 0,
+        "my_ulonglong": 0,
+        "my_ushort": 0,
+        "my_wchar": "\u0000",
+        "my_wstring": "",
+        "short_long_map": null,
+        "short_sequence": [],
+        "string_alias_unbounded_map": null,
+        "string_unbounded_map": null
+    }
+})";
