--- test/unittest/utils/json/DataType.json
@@ -1,268 +1,0 @@
-{
-    "complex_array": [
-        {
-            "bitmask_sequence": [],
-            "complex_union": {
-                "fourth": {
-                    "second": 0
-                }
-            },
-            "inner_union": {
-                "second": 0
-            },
-            "long_array": [
-                [
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0]
-                ],
-                [
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0]
-                ]
-            ],
-            "my_aliased_bounded_string": "",
-            "my_aliased_enum": {
-                "name": "A",
-                "value": 0
-            },
-            "my_aliased_struct": {
-                "my_bool": false,
-                "my_char": "\u0000",
-                "my_double": 0.0,
-                "my_float": 0.0,
-                "my_int8": 0,
-                "my_long": 0,
-                "my_longdouble": 0.0,
-                "my_longlong": 0,
-                "my_octet": 0,
-                "my_short": 0,
-                "my_uint8": 0,
-                "my_ulong": 0,
-                "my_ulonglong": 0,
-                "my_ushort": 0,
-                "my_wchar": "\u0000"
-            },
-            "my_bitmask": {
-                "active": [],
-                "binary": "00000000000000000000000000000000",
-                "value": 0
-            },
-            "my_bitset": {
-                "a": 0,
-                "b": false,
-                "c": 0,
-                "d": 0
-            },
-            "my_bool": false,
-            "my_bounded_string": "",
-            "my_bounded_wstring": "",
-            "my_char": "\u0000",
-            "my_double": 0.0,
-            "my_enum": {
-                "name": "A",
-                "value": 0
-            },
-            "my_float": 0.0,
-            "my_int8": 0,
-            "my_long": 0,
-            "my_longdouble": 0.0,
-            "my_longlong": 0,
-            "my_octet": 0,
-            "my_recursive_alias": {
-                "name": "A",
-                "value": 0
-            },
-            "my_short": 0,
-            "my_string": "",
-            "my_uint8": 0,
-            "my_ulong": 0,
-            "my_ulonglong": 0,
-            "my_ushort": 0,
-            "my_wchar": "\u0000",
-            "my_wstring": "",
-            "short_long_map": null,
-            "short_sequence": [],
-            "string_alias_unbounded_map": null,
-            "string_unbounded_map": null
-        },
-        {
-            "bitmask_sequence": [],
-            "complex_union": {
-                "fourth": {
-                    "second": 0
-                }
-            },
-            "inner_union": {
-                "second": 0
-            },
-            "long_array": [
-                [
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0]
-                ],
-                [
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0]
-                ]
-            ],
-            "my_aliased_bounded_string": "",
-            "my_aliased_enum": {
-                "name": "A",
-                "value": 0
-            },
-            "my_aliased_struct": {
-                "my_bool": false,
-                "my_char": "\u0000",
-                "my_double": 0.0,
-                "my_float": 0.0,
-                "my_int8": 0,
-                "my_long": 0,
-                "my_longdouble": 0.0,
-                "my_longlong": 0,
-                "my_octet": 0,
-                "my_short": 0,
-                "my_uint8": 0,
-                "my_ulong": 0,
-                "my_ulonglong": 0,
-                "my_ushort": 0,
-                "my_wchar": "\u0000"
-            },
-            "my_bitmask": {
-                "active": [],
-                "binary": "00000000000000000000000000000000",
-                "value": 0
-            },
-            "my_bitset": {
-                "a": 0,
-                "b": false,
-                "c": 0,
-                "d": 0
-            },
-            "my_bool": false,
-            "my_bounded_string": "",
-            "my_bounded_wstring": "",
-            "my_char": "\u0000",
-            "my_double": 0.0,
-            "my_enum": {
-                "name": "A",
-                "value": 0
-            },
-            "my_float": 0.0,
-            "my_int8": 0,
-            "my_long": 0,
-            "my_longdouble": 0.0,
-            "my_longlong": 0,
-            "my_octet": 0,
-            "my_recursive_alias": {
-                "name": "A",
-                "value": 0
-            },
-            "my_short": 0,
-            "my_string": "",
-            "my_uint8": 0,
-            "my_ulong": 0,
-            "my_ulonglong": 0,
-            "my_ushort": 0,
-            "my_wchar": "\u0000",
-            "my_wstring": "",
-            "short_long_map": null,
-            "short_sequence": [],
-            "string_alias_unbounded_map": null,
-            "string_unbounded_map": null
-        }
-    ],
-    "complex_map": null,
-    "complex_sequence": [],
-    "index": 0,
-    "inner_struct": {
-        "bitmask_sequence": [],
-        "complex_union": {
-            "fourth": {
-                "second": 0
-            }
-        },
-        "inner_union": {
-            "second": 0
-        },
-        "long_array": [
-            [
-                [0, 0, 0, 0],
-                [0, 0, 0, 0],
-                [0, 0, 0, 0]
-            ],
-            [
-                [0, 0, 0, 0],
-                [0, 0, 0, 0],
-                [0, 0, 0, 0]
-            ]
-        ],
-        "my_aliased_bounded_string": "",
-        "my_aliased_enum": {
-            "name": "A",
-            "value": 0
-        },
-        "my_aliased_struct": {
-            "my_bool": false,
-            "my_char": "\u0000",
-            "my_double": 0.0,
-            "my_float": 0.0,
-            "my_int8": 0,
-            "my_long": 0,
-            "my_longdouble": 0.0,
-            "my_longlong": 0,
-            "my_octet": 0,
-            "my_short": 0,
-            "my_uint8": 0,
-            "my_ulong": 0,
-            "my_ulonglong": 0,
-            "my_ushort": 0,
-            "my_wchar": "\u0000"
-        },
-        "my_bitmask": {
-            "active": [],
-            "binary": "00000000000000000000000000000000",
-            "value": 0
-        },
-        "my_bitset": {
-            "a": 0,
-            "b": false,
-            "c": 0,
-            "d": 0
-        },
-        "my_bool": false,
-        "my_bounded_string": "",
-        "my_bounded_wstring": "",
-        "my_char": "\u0000",
-        "my_double": 0.0,
-        "my_enum": {
-            "name": "A",
-            "value": 0
-        },
-        "my_float": 0.0,
-        "my_int8": 0,
-        "my_long": 0,
-        "my_longdouble": 0.0,
-        "my_longlong": 0,
-        "my_octet": 0,
-        "my_recursive_alias": {
-            "name": "A",
-            "value": 0
-        },
-        "my_short": 0,
-        "my_string": "",
-        "my_uint8": 0,
-        "my_ulong": 0,
-        "my_ulonglong": 0,
-        "my_ushort": 0,
-        "my_wchar": "\u0000",
-        "my_wstring": "",
-        "short_long_map": null,
-        "short_sequence": [],
-        "string_alias_unbounded_map": null,
-        "string_unbounded_map": null
-    }
-}
